// src/oncobridge-web/src/app/api/models/patient-record-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CancerStagingResponse } from '../models/cancer-staging-response';
import { CancerSurgicalProcedureResponse } from '../models/cancer-surgical-procedure-response';
import { PatientResponse } from '../models/patient-response';
import { PrimaryCancerDiagnosisResponse } from '../models/primary-cancer-diagnosis-response';
export interface PatientRecordResponse {
  cancerStagings: Array<CancerStagingResponse>;
  cancerSurgicalProcedures: Array<CancerSurgicalProcedureResponse>;
  patient: PatientResponse;
  primaryCancerDiagnoses: Array<PrimaryCancerDiagnosisResponse>;
}